Fixed a memory leak regarding HealthBars
Bug can also be seen on the showcase video on the repository (HH video).

Sometimes healthbars were floating in the game world randomly. Further investigations showed that, healthbars were actually not getting destroyed when the monsters are destroyed.

Issue was:
I was destroying the monsters in EntityManager class BUT, I wasn't destroying them as they being Monsters, instead, I was reaching to those monster variables as if they were Entity. Thus, Monster class' deconstructor wasn't getting called.  Which happens to be the one function that deals with destruction of healthbar pointers etc...